requirements.txt
django
south
django-mptt